+++ lib/libscarf/include/std__math.h
@@ -42,10 +42,15 @@ int		math__normalize(int num, const t_irange from, const t_irange to);
 /*
 ** < vector.c > */
 
-t_vec	vec__add(const t_vec *v1, const t_vec *v2);
-void	vec__add_assign(t_vec *v1, t_vec *v2);
 t_vec	vec__rotate(const t_vec *vec, double angle);
 void	vec__rotate_assign(t_vec *vec, double angle);
+/*
+** < vector_add.c > */
+
+t_vec	vec__add(const t_vec *v1, const t_vec *v2);
+void	vec__add_assign(t_vec *v1, t_vec *v2);
+t_vec	vec__sub(const t_vec *v1, const t_vec *v2);
+void	vec__sub_assign(t_vec *v1, t_vec *v2);
 /*
 ** < vector_mul.c > */
